Add Azure support and test

+++ src/providers/azure-blob/index.ts
@@ -1,4 +1,4 @@
-import { BlobDownloadResponseModel, BlobServiceClient, BlockBlobClient, ContainerClient } from "@azure/storage-blob"
+import { BlobDownloadResponseModel, BlobSASPermissions, BlobServiceClient, BlockBlobClient, ContainerClient, SASProtocol } from "@azure/storage-blob"
 import { CloudStorage } from "../../types/cloud"
 import { AzureConfig } from "../../types/config"
 import { FileMetadata } from "../../types/metadata"
@@ -10,10 +10,10 @@ export class AzureBlobService extends CloudStorage {
     constructor(config: AzureConfig) {
         super()
 
-        if (config.connectionString == null)
+        if (config.connectionString == null || config.connectionString.length === 0)
             throw new Error("Connection string must be provided")
 
-        if (config.container == null)
+        if (config.container == null || config.container.length === 0)
             throw new Error("Container parameter must be provided")
 
         this.client = BlobServiceClient.fromConnectionString(config.connectionString)
@@ -50,6 +50,9 @@ export class AzureBlobService extends CloudStorage {
 
     async downloadFile(key: string): Promise<BlobDownloadResponseModel> {
         try {
+            if (!await this.exists(key))
+                throw new Error("The specified key does not exist.")
+
             const blockBlobClient = this.bucket.getBlockBlobClient(key)
             return await blockBlobClient.download(0)
         } catch (error: any) {
@@ -59,6 +62,9 @@ export class AzureBlobService extends CloudStorage {
 
     async deleteFile(key: string): Promise<boolean> {
         try {
+            if (!await this.exists(key))
+                throw new Error("The specified key does not exist.")
+
             const blockBlobClient = this.bucket.getBlockBlobClient(key)
             await blockBlobClient.delete()
 
@@ -70,6 +76,9 @@ export class AzureBlobService extends CloudStorage {
 
     async getFile(key: string): Promise<FileMetadata> {
         try {
+            if (!await this.exists(key))
+                throw new Error("The specified key does not exist.")
+
             const blockBlobClient = this.bucket.getBlockBlobClient(key)
             const properties = await blockBlobClient.getProperties()
             const data = await this.downloadFile(key)
@@ -116,16 +125,27 @@ export class AzureBlobService extends CloudStorage {
 
     async getSignedUrl(key: string, expiresIn: number = 3600) {
         try {
-            const blockBlobClient: BlockBlobClient = this.bucket.getBlockBlobClient(key)
+            if (!await this.exists(key))
+                throw new Error("The specified key does not exist.")
 
-            const startsOn = new Date()
-            const expiresOn = new Date(startsOn)
+            const blockBlobClient: BlockBlobClient = this.bucket.getBlockBlobClient(key);
 
-            expiresOn.setSeconds(startsOn.getSeconds() + expiresIn)
+            const startsOn = new Date();
+            const expiresOn = new Date(startsOn);
+            expiresOn.setSeconds(startsOn.getSeconds() + expiresIn);
 
-            return await blockBlobClient.generateSasUrl({ startsOn, expiresOn })
+            // Specify the permissions
+            const permissions = BlobSASPermissions.parse("r"); // "r" for read access
+            const protocol = SASProtocol.HttpsAndHttp; // Optional: Restrict to HTTPS if needed
+
+            return await blockBlobClient.generateSasUrl({
+                permissions,
+                startsOn,
+                expiresOn,
+                protocol, // Optional
+            });
         } catch (error: any) {
-            throw new Error(`Failed to generate signed URL: ${error.message}`)
+            throw new Error(`Failed to generate signed URL: ${error.message}`);
         }
     }
 }
